renamed randomRDD to rand and import as random
